// src/app/shared/models/archetype-query-builder/referencemodel-type.enum.ts
export enum ReferenceModelType {
  Access_control_settings = 'ACCESS_CONTROL_SETTINGS',
  Access_group_ref = 'ACCESS_GROUP_REF',
  Action = 'ACTION',
  Activity = 'ACTIVITY',
  Admin_entry = 'ADMIN_ENTRY',
  Any = 'ANY',
  Archetype_hrid = 'ARCHETYPE_HRID',
  Archetype_id = 'ARCHETYPE_ID',
  Archetyped = 'ARCHETYPED',
  Array = 'ARRAY',
  Attestation = 'ATTESTATION',
  Audit_details = 'AUDIT_DETAILS',
  Authored_resource = 'AUTHORED_RESOURCE',
  Boolean = 'BOOLEAN',
  Byte = 'BYTE',
  Care_entry = 'CARE_ENTRY',
  Character = 'CHARACTER',
  Cluster = 'CLUSTER',
  Code_phrase = 'CODE_PHRASE',
  Composition = 'COMPOSITION',
  Container = 'CONTAINER',
  Content_item = 'CONTENT_ITEM',
  Contribution = 'CONTRIBUTION',
  Data_structure = 'DATA_STRUCTURE',
  Data_value = 'DATA_VALUE',
  Double = 'DOUBLE',
  Dv_absolute_quantity = 'DV_ABSOLUTE_QUANTITY',
  Dv_amount = 'DV_AMOUNT',
  Dv_boolean = 'DV_BOOLEAN',
  Dv_coded_text = 'DV_CODED_TEXT',
  Dv_count = 'DV_COUNT',
  Dv_date = 'DV_DATE',
  Dv_date_time = 'DV_DATE_TIME',
  Dv_duration = 'DV_DURATION',
  Dv_ehr_uri = 'DV_EHR_URI',
  Dv_encapsulated = 'DV_ENCAPSULATED',
  Dv_general_time_specification = 'DV_GENERAL_TIME_SPECIFICATION',
  Dv_identifier = 'DV_IDENTIFIER',
  Dv_interval = 'DV_INTERVAL',
  Dv_multimedia = 'DV_MULTIMEDIA',
  Dv_ordered = 'DV_ORDERED',
  Dv_ordinal = 'DV_ORDINAL',
  Dv_paragraph = 'DV_PARAGRAPH',
  Dv_parsable = 'DV_PARSABLE',
  Dv_periodic_time_specification = 'DV_PERIODIC_TIME_SPECIFICATION',
  Dv_proportion = 'DV_PROPORTION',
  Dv_quantified = 'DV_QUANTIFIED',
  Dv_quantity = 'DV_QUANTITY',
  Dv_state = 'DV_STATE',
  Dv_temporal = 'DV_TEMPORAL',
  Dv_text = 'DV_TEXT',
  Dv_time = 'DV_TIME',
  Dv_time_specification = 'DV_TIME_SPECIFICATION',
  Dv_uri = 'DV_URI',
  Ehr = 'EHR',
  Ehr_access = 'EHR_ACCESS',
  Ehr_status = 'EHR_STATUS',
  Element = 'ELEMENT',
  Entry = 'ENTRY',
  Evaluation = 'EVALUATION',
  Event = 'EVENT',
  Event_context = 'EVENT_CONTEXT',
  Feeder_audit = 'FEEDER_AUDIT',
  Feeder_audit_details = 'FEEDER_AUDIT_DETAILS',
  Folder = 'FOLDER',
  Generic_entry = 'GENERIC_ENTRY',
  Generic_id = 'GENERIC_ID',
  Hash = 'HASH',
  Hier_object_id = 'HIER_OBJECT_ID',
  History = 'HISTORY',
  Imported_version = 'IMPORTED_VERSION',
  Instruction = 'INSTRUCTION',
  Instruction_details = 'INSTRUCTION_DETAILS',
  Integer = 'INTEGER',
  Integer64 = 'INTEGER64',
  Internet_id = 'INTERNET_ID',
  Interval = 'INTERVAL',
  Interval_event = 'INTERVAL_EVENT',
  Ism_transition = 'ISM_TRANSITION',
  Iso_oid = 'ISO_OID',
  Item = 'ITEM',
  Item_list = 'ITEM_LIST',
  Item_single = 'ITEM_SINGLE',
  Item_structure = 'ITEM_STRUCTURE',
  Item_table = 'ITEM_TABLE',
  Item_tree = 'ITEM_TREE',
  Link = 'LINK',
  List = 'LIST',
  Locatable = 'LOCATABLE',
  Locatable_ref = 'LOCATABLE_REF',
  Long = 'LONG',
  Numeric = 'NUMERIC',
  Object_id = 'OBJECT_ID',
  Object_ref = 'OBJECT_REF',
  Object_version_id = 'OBJECT_VERSION_ID',
  Observation = 'OBSERVATION',
  Octet = 'OCTET',
  Ordered = 'ORDERED',
  Ordered_numeric = 'ORDERED_NUMERIC',
  Original_version = 'ORIGINAL_VERSION',
  Participation = 'PARTICIPATION',
  Party_identified = 'PARTY_IDENTIFIED',
  Party_proxy = 'PARTY_PROXY',
  Party_ref = 'PARTY_REF',
  Party_related = 'PARTY_RELATED',
  Party_self = 'PARTY_SELF',
  Pathable = 'PATHABLE',
  Point_event = 'POINT_EVENT',
  Proportion_kind = 'PROPORTION_KIND',
  Real = 'REAL',
  Reference_range = 'REFERENCE_RANGE',
  Resource_description = 'RESOURCE_DESCRIPTION',
  Resource_description_item = 'RESOURCE_DESCRIPTION_ITEM',
  Revision_history = 'REVISION_HISTORY',
  Revision_history_item = 'REVISION_HISTORY_ITEM',
  Section = 'SECTION',
  Set = 'SET',
  String = 'STRING',
  Template_id = 'TEMPLATE_ID',
  Temporal_accessor = 'TEMPORAL_ACCESSOR',
  Temporal_amount = 'TEMPORAL_AMOUNT',
  Term_mapping = 'TERM_MAPPING',
  Terminology_code = 'TERMINOLOGY_CODE',
  Terminology_id = 'TERMINOLOGY_ID',
  Terminology_term = 'TERMINOLOGY_TERM',
  Translation_details = 'TRANSLATION_DETAILS',
  Uid = 'UID',
  Uid_based_id = 'UID_BASED_ID',
  Uri = 'URI',
  Uuid = 'UUID',
  Validity_kind = 'VALIDITY_KIND',
  Version = 'VERSION',
  Version_status = 'VERSION_STATUS',
  Version_tree_id = 'VERSION_TREE_ID',
  Versioned_object = 'VERSIONED_OBJECT',
}